Split out gui code into gui sub-project
